Portfolio polish: bilingual README, screenshots, license, release APK + crash fix
- Fix IllegalFormatConversionException crash in RunFragment.checkVoiceMilestone:
  the km milestone was an Int but tts_voice_milestone uses a %.0f format, so the
  app crashed on every run that crossed a voice milestone (≥1 km, voice on by
  default). Now passes a Float. Verified on a 2 km emulator run (Trikala).
- Rewrite README.md as a portfolio-grade English page (badges, screenshots,
  download, build/run, architecture) + add Greek README.el.md.
- Add MIT LICENSE.
- Add docs/screenshots/ (real demo runs recorded in Trikala, Greece).
- Add signed release APK at release/FitnessUltra-v1.0.apk; release signing reads
  a git-ignored keystore.properties (keystore + passwords never committed).
- Commit fragment_offline_maps.xml (was untracked — a fresh clone would not build).

+++ app/src/main/java/com/fitnessultra/ui/run/RunFragment.kt
@@ -422,7 +422,8 @@ class RunFragment : Fragment() {
         val milestonePassed = (distanceKm / freqKm).toInt()
         if (milestonePassed > lastVoiceKm) {
             lastVoiceKm = milestonePassed
-            val kmAnnounced = milestonePassed * freqKm
+            // Must be Float: tts_voice_milestone uses a %.0f format specifier.
+            val kmAnnounced = (milestonePassed * freqKm).toFloat()
             val durationMs = viewModel.timeRunInMillis.value ?: 0L
             val paceStr = TrackingUtils.calculatePace(kmAnnounced * 1000f, durationMs, context = requireContext())
             speakTts(getTtsString(R.string.tts_voice_milestone, kmAnnounced, paceStr))
